Add test tool for fuse fs

+++ ext2fs.c
@@ -360,7 +360,8 @@ int ext2_main(int argc, char *argv[], void (*mounted) (void))
 	}
 	fuse_opt_free_args(&args);
 
-	rmdir (mountpoint);
+	if (mounted)
+	  rmdir (mountpoint);
 	
 	return err ? 1 : 0;
 }
